+++ src/ArduinoJson/Serialization/measure.hpp
@@ -5,6 +5,7 @@
 #pragma once
 
 #include <ArduinoJson/Serialization/Writers/DummyWriter.hpp>
+#include <ArduinoJson/Variant/VariantFunctions.hpp>
 
 namespace ARDUINOJSON_NAMESPACE {
 
@@ -12,7 +13,7 @@ template <template <typename> class TSerializer>
 size_t measure(VariantConstRef source) {
   DummyWriter dp;
   TSerializer<DummyWriter> serializer(dp);
-  return source.accept(serializer);
+  return variantAccept(getData(source), serializer);
 }
 
 }  // namespace ARDUINOJSON_NAMESPACE