+++ CHANGELOG.md
@@ -7,6 +7,7 @@ HEAD
 * Add `JsonVariant::link()` (issue #1343)
 * Fix `9.22337e+18 is outside the range of representable values of type 'long'`
 * Fix comparison operators for `JsonArray`, `JsonArrayConst`, `JsonObject`, and `JsonObjectConst`
+* Remove undocumented `accept()` functions
 
 v6.19.4 (2022-04-05)
 -------